Add Price property to book-related DTOs and entity
This commit introduces a new `Price` property to the `CreateBookDto`, `ResultBookDto`, `UpdateBookDto`, and `Book` entity, allowing for the management of book prices. A migration (`20250603150651_bookUpdPrice.cs`) has been created to update the database schema with a new `Price` column in the `Books` table.

Additionally, the project file has been updated to include a reference to the `Booksaw.Dto` project, and the `Program.cs` file has been modified to add HTTP client services. Several Razor view files have been enhanced to display books and their prices, with new components (`_BillboardComponent` and `_FeaturedBooksComponent`) created for fetching and displaying book lists from an API. Overall, these changes improve the user interface and functionality of the application.